rearranged timer to frameratecontroller
