--- myRenderer3/Timer.cpp
@@ -1,8 +1,0 @@
-#include "timer.h"
-
-Timer nFrequency;
-
-void initialiseTimer()
-{
-    QueryPerformanceFrequency(&nFrequency); //to be called only once. not per frame
-}
