--- myRenderer3/Timer.h
@@ -1,8 +1,0 @@
-#pragma once
-#include <Windows.h>
-
-typedef LARGE_INTEGER Timer;
-
-extern Timer nFrequency;
-
-void initialiseTimer();
