+++ myRenderer3/myRenderer3.vcxproj
@@ -130,7 +130,6 @@
     <ClCompile Include="getKeyInput.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="runMainLoop.cpp" />
-    <ClCompile Include="timer.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="fileManager.h" />
@@ -140,7 +139,6 @@
     <ClInclude Include="frameRateController.h" />
     <ClInclude Include="getKeyInput.h" />
     <ClInclude Include="runMainLoop.h" />
-    <ClInclude Include="timer.h" />
   </ItemGroup>
   <ItemGroup>
     <Text Include="config.txt" />
